Added Babel, Updated linux pygame, and dropped futures
Babel adds the pybabel command to extract the messages from the code.
Flask-Babel integrates Babel with Flask's request objects.
futures is part of the standard lib for python3 now, so doesn't want to
be installed separately.

I needed to update pygame to get it to launch on my dev computer, though
I can't try it on a pi right now.